Delete runtime library and the pass
Delete all files. The following commit will import files from llvm
project, and then the following of the following commit will recreate
files deleted by this commit.

This might seem unnecessary, but this is to simplify the process of
rebasing this project to the latest LLVM source code later.

While this breaks building and testing for a few commits, it simplifies
the process of rebasing.

Signed-off-by: Hyeonggon Yoo <42.hyeyoo@gmail.com>