PreciseLeakSanitizer: Use own allocator and integrate LSan on it
Currently, LSan is linked together when PLSan is enabled.
Use own allocator and integrate LSan on it.

PLSan allocator is a typical allocator for a sanitizer, with some
functions needed to run LSan on it.

LSan is initialized in __plsan_init() constructor, and then invoked when
the process exits.

At some point we need to move some code into the allocator instead of
inserting function calls.

Signed-off-by: Hyeonggon Yoo <42.hyeyoo@gmail.com>