compiler-rt: Fix not decrementing the reference count in some cases
The commit c67230c86bdb incorrectly implemented UpdateReference()
function, it does not decrease the reference count when the previous
value of a pointer variable was a valid heap pointer.

Fix this by correcting the parameter's type, and drop constness which is
not that much useful here. This fixes the testcase testcases/c/null-3.c.

Signed-off-by: Hyeonggon Yoo <42.hyeyoo@gmail.com>