compiler-rt: Add precise leak report in lsan
When the program terminates, lsan detects the leak.
If plsan knows the location where it is lost, it outputs that location.

PLSan reports only leaks detected by LSan to avoid false positives.
However, there are some type of leaks only PLSan can detect. An option to report
memory leaks more aggressively (even when LSan did not detect) can be added later.

For now, we modify LSan code to call plsan handler function if PLSan is enabled
(using weak symbols, see __plsan_is_enabled() for details). We might improve
this in the further to avoid modifying LSan as much as possible.

While integrating LSan, the init code had a few problems:
- it called InitializedThreads(), leading to crashes

... also some pieces of init code are borrowed from LSan.

Also, there was a leak because InternalAlloc()-ed memory was not properly freed.
Fix it by free()-ing when it is not used anymore.

Now that PlsanHandler is not needed anymore, drop PlsanHandler.{cpp,h} completely.
Also delete backward.h which is not used anymore.