compiler-rt: Do not mess up with metadata, when allocated from the secondary
When an object is allocated from the secondary allocator (because the
size is big), it does not correctly get metadata because I missed
GetBlockBegin() when getting the metadata.

Fix this by calling GetBlockBegin() first before calling
allocator.GetMetaData(). For just slight optimization, remove checking
if it is allocated from the primary allocator, because it is already
done in allocator.GetBlockBegin().

Signed-off-by: Hyeonggon Yoo <42.hyeyoo@gmail.com>