compiler-rt: optimize Pointer-Is-Mine checking for the primary allocator
When profiling programs with the Linux perf tool, approximately 30%
of the profile is consumed by checking if an address has been allocated
by the allocator (PointerIsMine()), and about 11% of the total profile is
dedicated to verifying if an address is allocated by the primary allocator.

Therefore, optimizing the process to quickly determine if an address
belongs to the primary allocator's address space is worthwhile.

The primary allocator operates within a distinct address space for
memory allocation. To enhance efficiency, avoid calling functions
like SpaceBeg(), SpaceEnd(), and instead utilize a constant address space
for the primary allocator.

This check can later be inserted into the program as inlined.
(i.e. check address first before calling __plsan_store())

Signed-off-by: Hyeonggon Yoo <42.hyeyoo@gmail.com>