Fix: fix builtin alloca error
If there is builtin alloca instruction, then that instruction doesn't
come to top of IR entry block.

If the builtin alloca instruction is not executed, it is possiable to
call free_local_variable about garbage value.

It can be fixed with one is_executed boolean variable per one builtin
alloca instruction to decide that it is actually allocated.