compiler-rt: Fix incorrect optimization in PtrIsAllocatedFromPlsan()
The commit 30a6aaf66c5 ("compiler-rt: optimize Pointer-Is-Mine checking
for the primary allocator") is fundamentally wrong, as not being
allocated from the primary allocator does not mean it's not allocated
from the secondary.

Unless we deal with the secondary allocator's address space, do not
determine it by address space.

By the way, PtrIsAllocatedFromPlsan() is unnecessarily called in
UpdateReference(), because GetMetadata() is already called in
IncRefCount() and DecRefCount(). Thus drop them for better performance.

Signed-off-by: Hyeonggon Yoo <42.hyeyoo@gmail.com>